pyproject.toml
[project]
name = "speech2text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.34.2",
    "editdistance>=0.8.1",
    "faster-whisper>=1.0.3",
    "fastparquet>=2024.5.0",
    "google-generativeai>=0.7.2",
    "jupyter>=1.1.1",
    "langchain>=0.3.1",
    "langchain-google-genai>=2.0.0",
    "librosa>=0.10.2.post1",
    "matplotlib>=3.9.2",
    "num2words>=0.5.13",
    "numba==0.60.0",
    "numpy>=1.22",
    "openai>=1.50.2",
    "openai-whisper>=20240927",
    "pandarallel>=1.6.5",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pyarrow>=17.0.0",
    "pytelegrambotapi>=4.23.0",
    "sentencepiece>=0.2.0",
    "soundfile>=0.12.1",
    "torch-audiomentations>=0.11.1",
    "torch-lr-finder>=0.2.2",
    "torchmetrics>=1.4.2",
    "tqdm>=4.66.5",
    "transformers>=4.45.1",
    "whisperx>=3.1.2",
    "python-dotenv>=1.0.1",
    "pydub>=0.25.1",
]
